dfu: Fix uploading large firmware files
Fixes split into more than 0xff chunks were being truncated as uint8_t was the
wrong index type.

+++ plugins/dfu/dfu-target-private.h
@@ -33,11 +33,11 @@ DfuTarget	*dfu_target_new				(DfuDevice	*device,
 							 GUsbInterface	*iface);
 
 GBytes		*dfu_target_upload_chunk		(DfuTarget	*target,
-							 guint8		 index,
+							 guint16	 index,
 							 GCancellable	*cancellable,
 							 GError		**error);
 gboolean	 dfu_target_download_chunk		(DfuTarget	*target,
-							 guint8		 index,
+							 guint16	 index,
 							 GBytes		*bytes,
 							 GCancellable	*cancellable,
 							 GError		**error);
